+++ types/langmap/index.d.ts
@@ -3,15 +3,8 @@
 // Definitions by: Gábor Balogh <https://github.com/grabofus>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
-export interface Language {
-    englishName: string;
-    nativeName: string;
-}
-
-export interface LanguageMappingList {
-    [language: string]: Language;
-}
+import { LanguageMappingList } from "./types";
 
 declare const langmap: LanguageMappingList;
 
-export default langmap;
+export = langmap;
