+++ types/langmap/langmap-tests.ts
@@ -1,3 +1,11 @@
-import langmap from 'langmap';
+import langmap1 from 'langmap';
+import * as langmap2 from 'langmap';
 
-langmap['en-US'] // { nativeName: "English (US)", englishName: "English (US)" }
+const val1 = langmap1['en-US']; // { nativeName: string, englishName: string }
+const val2 = langmap2['en-US']; // { nativeName: string, englishName: string }
+
+val1.englishName;
+val1.nativeName;
+
+val2.englishName;
+val2.nativeName;
