+++ types/langmap/tsconfig.json
@@ -4,6 +4,7 @@
         "lib": [
             "es6"
         ],
+        "esModuleInterop": true,
         "noImplicitAny": true,
         "noImplicitThis": true,
         "strictFunctionTypes": true,
@@ -18,6 +19,7 @@
     },
     "files": [
         "index.d.ts",
-        "langmap-tests.ts"
+        "langmap-tests.ts",
+        "types.d.ts"
     ]
 }
