+++ types/langmap/types.d.ts
@@ -1,0 +1,8 @@
+export interface Language {
+    englishName: string;
+    nativeName: string;
+}
+
+export interface LanguageMappingList {
+    [language: string]: Language;
+}
